pattern-formatter/backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-docx==1.1.0
docxcompose==1.4.0
gunicorn==21.2.0
pywin32>=300; sys_platform == 'win32'
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
lxml==5.1.0
Pillow==10.2.0
